refactor: extract profile launch runner from main process
Made-with: Cursor
